condense platform-specific logic into one file per platform - change retryIfInternalServerError to retryOnInternalServerErrorTimes to allow multiple times/infinite retry on internal server error - remove redundant base64 code
