+++ src/androidMain/kotlin/com/adamratzman/spotify/utils/PlatformUtils.kt
@@ -1,11 +1,21 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
 package com.adamratzman.spotify.utils
 
+import android.annotation.SuppressLint
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.util.Base64
 import java.io.ByteArrayOutputStream
 import java.net.URL
+import java.net.URLEncoder
+import java.text.SimpleDateFormat
+import java.util.Date
+
+internal actual fun String.encodeUrl() = URLEncoder.encode(this, "UTF-8")!!
+
+@SuppressLint("SimpleDateFormat")
+internal actual fun formatDate(format: String, date: Long): String {
+    return SimpleDateFormat(format).format(Date(date))
+}
 
 internal actual fun encodeBufferedImageToBase64String(image: BufferedImage): String {
     val byteArrayOutputStream = ByteArrayOutputStream()
@@ -23,3 +33,16 @@ internal actual fun convertUrlPathToBufferedImage(url: String): BufferedImage {
         BitmapFactory.decodeStream(inputStream)
     }
 }
+
+/**
+ * Actual platform that this program is run on.
+ */
+public actual val currentApiPlatform: Platform = Platform.ANDROID
+
+public actual typealias ConcurrentHashMap<K, V> = java.util.concurrent.ConcurrentHashMap<K, V>
+
+public actual typealias BufferedImage = Bitmap // TODO
+
+public actual typealias File = java.io.File
+
+public actual fun <K, V> ConcurrentHashMap<K, V>.asList(): List<Pair<K, V>> = toList()
