--- src/androidMain/kotlin/com/adamratzman/spotify/utils/Types.kt
@@ -1,12 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.utils
-
-import android.graphics.Bitmap
-
-public actual typealias ConcurrentHashMap<K, V> = java.util.concurrent.ConcurrentHashMap<K, V>
-
-public actual typealias BufferedImage = Bitmap // TODO
-
-public actual typealias File = java.io.File
-
-public actual fun <K, V> ConcurrentHashMap<K, V>.asList(): List<Pair<K, V>> = toList()
