+++ src/commonMain/kotlin/com.adamratzman.spotify/Builder.kt
@@ -863,7 +863,8 @@ public class SpotifyUserAuthorization(
  * @param requiredScopes Scopes that your application requires to function (only applicable to [SpotifyClientApi] and [SpotifyImplicitGrantApi]).
  * @param proxyBaseUrl Provide if you have a proxy base URL that you would like to use instead of the Spotify API base
  * (https://api.spotify.com/v1).
- * @param retryIfInternalServerError Whether to retry once if an internal server error (500..599) has been received
+ * @param retryOnInternalServerErrorTimes Whether and how often to retry once if an internal server error (500..599) has been received. Set to 0
+ * to avoid retrying at all, or set to null to keep retrying until success.
  *
  */
 public data class SpotifyApiOptions(
@@ -881,5 +882,5 @@ public data class SpotifyApiOptions(
     public var onTokenRefresh: (suspend (GenericSpotifyApi) -> Unit)? = null,
     public var requiredScopes: List<SpotifyScope>? = null,
     public var proxyBaseUrl: String? = null,
-    public var retryIfInternalServerError: Boolean = true
+    public var retryOnInternalServerErrorTimes: Int? = 5,
 )
