+++ src/commonMain/kotlin/com.adamratzman.spotify/endpoints/client/ClientEpisodeApi.kt
@@ -5,7 +5,7 @@ import com.adamratzman.spotify.GenericSpotifyApi
 import com.adamratzman.spotify.SpotifyException.BadRequestException
 import com.adamratzman.spotify.SpotifyScope
 import com.adamratzman.spotify.endpoints.public.EpisodeApi
-import com.adamratzman.spotify.http.encodeUrl
+import com.adamratzman.spotify.utils.encodeUrl
 import com.adamratzman.spotify.models.Episode
 import com.adamratzman.spotify.models.EpisodeList
 import com.adamratzman.spotify.models.EpisodeUri
