+++ src/commonMain/kotlin/com.adamratzman.spotify/endpoints/public/AlbumApi.kt
@@ -4,7 +4,7 @@ package com.adamratzman.spotify.endpoints.public
 import com.adamratzman.spotify.GenericSpotifyApi
 import com.adamratzman.spotify.SpotifyException.BadRequestException
 import com.adamratzman.spotify.http.SpotifyEndpoint
-import com.adamratzman.spotify.http.encodeUrl
+import com.adamratzman.spotify.utils.encodeUrl
 import com.adamratzman.spotify.models.Album
 import com.adamratzman.spotify.models.AlbumUri
 import com.adamratzman.spotify.models.AlbumsResponse
