+++ src/commonMain/kotlin/com.adamratzman.spotify/endpoints/public/ShowApi.kt
@@ -6,7 +6,7 @@ import com.adamratzman.spotify.SpotifyAppApi
 import com.adamratzman.spotify.SpotifyException.BadRequestException
 import com.adamratzman.spotify.SpotifyScope
 import com.adamratzman.spotify.http.SpotifyEndpoint
-import com.adamratzman.spotify.http.encodeUrl
+import com.adamratzman.spotify.utils.encodeUrl
 import com.adamratzman.spotify.models.PagingObject
 import com.adamratzman.spotify.models.Show
 import com.adamratzman.spotify.models.ShowList
