+++ src/commonMain/kotlin/com.adamratzman.spotify/endpoints/public/UserApi.kt
@@ -3,7 +3,7 @@ package com.adamratzman.spotify.endpoints.public
 
 import com.adamratzman.spotify.GenericSpotifyApi
 import com.adamratzman.spotify.http.SpotifyEndpoint
-import com.adamratzman.spotify.http.encodeUrl
+import com.adamratzman.spotify.utils.encodeUrl
 import com.adamratzman.spotify.models.SpotifyPublicUser
 import com.adamratzman.spotify.models.UserUri
 import com.adamratzman.spotify.models.serialization.toObject
