+++ src/commonMain/kotlin/com.adamratzman.spotify/utils/Concurrency.kt
@@ -4,8 +4,8 @@ package com.adamratzman.spotify.utils
 /**
  * Represents a unit of time
  */
-public expect enum class TimeUnit {
-    MILLISECONDS, SECONDS;
+public enum class TimeUnit(public val multiplier: Long) {
+    MILLISECONDS(1), SECONDS(1000), MINUTES(60000);
 
-    public fun toMillis(duration: Long): Long
+    public fun toMillis(duration: Long): Long = duration * multiplier
 }
