--- src/desktopMain/kotlin/com/adamratzman/spotify/utils/Types.kt
@@ -1,8 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.utils
-
-public actual typealias BufferedImage = String
-
-public actual typealias ConcurrentHashMap<K, V> = HashMap<K, V>
-
-public actual fun <K, V> ConcurrentHashMap<K, V>.asList(): List<Pair<K, V>> = toList()
