--- src/jsMain/kotlin/com/adamratzman/spotify/http/Endpoints.kt
@@ -1,11 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.http
-
-import com.adamratzman.spotify.utils.encodeToBase64
-import io.ktor.http.encodeURLQueryComponent
-
-internal actual fun String.base64ByteEncode(): String {
-    return this.encodeToBase64()
-}
-
-internal actual fun String.encodeUrl() = encodeURLQueryComponent()
