--- src/jsMain/kotlin/com/adamratzman/spotify/utils/Platform.kt
@@ -1,7 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.utils
-
-/**
- * Actual platform that this program is run on.
- */
-public actual val currentApiPlatform: Platform = Platform.JS
