--- src/jsMain/kotlin/com/adamratzman/spotify/utils/Types.kt
@@ -1,10 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.utils
-
-import org.w3c.files.File
-
-public actual typealias BufferedImage = File
-
-public actual typealias ConcurrentHashMap<K, V> = HashMap<K, V>
-
-public actual fun <K, V> ConcurrentHashMap<K, V>.asList(): List<Pair<K, V>> = toList()
