--- src/jvmMain/kotlin/com/adamratzman/spotify/utils/Concurrency.kt
@@ -1,7 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.utils
-
-/**
- * Represents a unit of time
- */
-public actual typealias TimeUnit = java.util.concurrent.TimeUnit
