--- src/jvmMain/kotlin/com/adamratzman/spotify/utils/Types.kt
@@ -1,10 +1,0 @@
-/* Spotify Web API, Kotlin Wrapper; MIT License, 2017-2021; Original author: Adam Ratzman */
-package com.adamratzman.spotify.utils
-
-public actual typealias ConcurrentHashMap<K, V> = java.util.concurrent.ConcurrentHashMap<K, V>
-
-public actual typealias BufferedImage = java.awt.image.BufferedImage
-
-public actual typealias File = java.io.File
-
-public actual fun <K, V> ConcurrentHashMap<K, V>.asList(): List<Pair<K, V>> = toList()
