+++ src/mingwX64Main/kotlin/com.adamratzman.spotify.utils/DesktopPlatformUtils.kt
